+++ Ryn.Core/Source/Clock.cpp
@@ -7,10 +7,10 @@ namespace Ryn::Core
         _start = Current();
     }
 
-    f32 Clock::Restart()
+    f64 Clock::Restart()
     {
-        f32 current = Current();
-        f32 elapsed = current - _start;
+        f64 current = Current();
+        f64 elapsed = current - _start;
         _start = current;
         return elapsed;
     }
